Refactor User model to change UserID type from uint64 to uint and enhance User function to include queue existence and status retrieval

+++ backend/src/models/queue.go
@@ -9,7 +9,7 @@ import (
 
 type Queue struct {
 	ID           uint64
-	UserID       uint64
+	UserID       uint
 	Status       string
 	AttemptCount uint8
 	LastError    sql.NullString
